services/alerts-normalizer-svc/requirements.txt
faust-streaming==0.10.10
fastapi==0.109.0
uvicorn[standard]==0.24.0.post1
httpx==0.25.2
python-dotenv==1.0.0
pydantic==1.10.14
loguru==0.7.2
